TS mejora: Estructura para types e interfaces.
