+++ src/types/components/docs/index.ts
@@ -1,0 +1,23 @@
+export interface RadioOption {
+  label: string
+  value: string
+}
+
+export interface RadioProps {
+  name: string
+  options: RadioOption[]
+  defaultValue?: string
+}
+
+export interface RadioOptionProps extends RadioOption {
+  name: string
+  selected: boolean
+  onClick: () => void
+}
+
+
+export interface ColorProps {
+  name: string
+  hex: string
+  description?: string
+}
